basic game working before offset

+++ src/main.py
@@ -1,8 +1,7 @@
-from random import randint
-import pygame
-from os.path import join
-
 from settings import *
+
+from random import randint
+from os.path import join
 
 
 class Timer:
@@ -40,7 +39,7 @@ class SnakeSegment(pygame.sprite.Sprite):
         super().__init__(groups)
         self.pos = pos
         self.image = pygame.Surface((CELL_SIZE, CELL_SIZE))
-        self.image.fill(pygame.Color(43, 51, 24))
+        self.image.fill(DARK_GREEN)
         self.rect = self.image.get_frect(topleft=self.pos)
 
 
@@ -56,11 +55,15 @@ class Snake():
                 4 * CELL_SIZE, 9 * CELL_SIZE), all_sprites)
         ]
         self.direction = pygame.Vector2(1, 0)
+        self.add_segment = False
 
     def update(self):
         self.body.insert(
             0, (SnakeSegment(self.body[0].pos + self.direction * CELL_SIZE, self.all_sprites)))
-        self.body.pop().kill()
+        if self.add_segment:
+            self.add_segment = False
+        else:
+            self.body.pop().kill()
 
 
 class Game:
@@ -72,7 +75,8 @@ class Game:
         pygame.mouse.set_visible(False)
         self.running = True
 
-        self.timer = Timer(200)
+        self.update_interval = 200
+        self.timer = Timer(self.update_interval)
 
         # imports
         self.load_data()
@@ -89,7 +93,6 @@ class Game:
 
     def get_input(self):
         keys = pygame.key.get_pressed()
-        print(self.snake.direction)
         if keys[pygame.K_RIGHT] and self.snake.direction.x != -1:
             self.snake.direction = pygame.Vector2(1, 0)
         if keys[pygame.K_LEFT] and self.snake.direction.x != 1:
@@ -102,6 +105,25 @@ class Game:
     def check_food_collision(self):
         if self.snake.body[0].rect.colliderect(self.food.rect):
             self.food.set_new_pos()
+            self.snake.add_segment = True
+            self.update_interval -= 2
+            self.timer.interval = self.update_interval
+
+    def check_edge_collision(self):
+        if self.snake.body[0].rect.x == (CELL_COUNT * CELL_SIZE) or self.snake.body[0].rect.x <= -1:
+            self.game_over()
+        if self.snake.body[0].rect.y == CELL_COUNT * CELL_SIZE or self.snake.body[0].rect.y <= -1:
+            self.game_over()
+
+    def check_tail_collision(self):
+        tail = self.snake.body[1:]
+        head = self.snake.body[0]
+        for segment in tail:
+            if head.rect.colliderect(segment.rect):
+                self.game_over()
+
+    def game_over(self):
+        self.__init__()
 
     def run(self):
         while self.running:
@@ -114,14 +136,19 @@ class Game:
 
             # logic
             self.check_food_collision()
+            self.check_edge_collision()
+            self.check_tail_collision()
 
             if self.timer.update_event():
                 self.snake.update()
             self.all_sprites.update()
 
             # draw
-            self.display_surf.fill(pygame.Color(173, 204, 96))
+            self.display_surf.fill(LIGHT_GREEN)
             self.all_sprites.draw(self.display_surf)
+
+            # pygame.draw.rect(self.display_surf, DARK_GREEN, (OFFSET-5,
+            #                  OFFSET-5, CELL_SIZE*CELL_COUNT, CELL_SIZE*CELL_COUNT+10), 5, 2)
 
             pygame.display.update()
